requirements.txt
requests==2.31.0
Pillow==10.2.0 